ehanced location search: - search request are now made using a map boundary - search results are only computed for the map boundary - the number of results is adopted to the results in the visible range - added a double-buffering for the search result markers - added a search query option for the search results: /radius/<lat>/<lon>/<radius>

+++ source/net/yacy/kelondro/data/meta/URIMetadata.java
@@ -9,12 +9,12 @@
  *  modify it under the terms of the GNU Lesser General Public
  *  License as published by the Free Software Foundation; either
  *  version 2.1 of the License, or (at your option) any later version.
- *  
+ *
  *  This library is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *  Lesser General Public License for more details.
- *  
+ *
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program in the file lgpl21.txt
  *  If not, see <http://www.gnu.org/licenses/>.
@@ -31,17 +31,17 @@ import net.yacy.kelondro.rwi.Reference;
 public interface URIMetadata extends URIReference {
 
     public String dc_title();
-    
+
     public String dc_creator();
-    
+
     public String dc_publisher();
-    
+
     public String dc_subject();
-    
-    public float lat();
-    
-    public float lon();
-    
+
+    public double lat();
+
+    public double lon();
+
     public long ranking();
 
     public Date loaddate();
@@ -71,13 +71,13 @@ public interface URIMetadata extends URIReference {
     public int lvideo();
 
     public int lapp();
-    
+
     public String snippet();
 
     public Reference word();
 
     public boolean isOlder(final URIMetadata other);
-    
+
     public String toString(final String snippet);
 
 }
